Shoehorn OAuth support into imapdriver_login_sasl.
This calls mailimap_oauth2_authenticate instead of
mailimap_authenticate if auth_type == "OAUTH2".

This means that mailstorage_connect can be passed a record with
auth_type == "OAUTH2" and it will use OAuth instead of SASL.

It takes the access_token from the password field.

This also adds IMAP_AUTH_TYPE_OAUTH2 as a constant.  This is not used
inside libetpan, but it makes it easier for MailCore.