Support mb == NULL as a parameter to imapdriver_select_folder.
In this case, we throw away the current imap_mailbox, and return
success.  This can be used to reset the state so that the next
call to imapdriver_select_folder will perform the
imap_flags_store_process and mailimap_select calls even if we're
not actually switching folders.

This allows us to get the mailimap_selection_info data (sel_uidnext
etc) from the server rather than using the cached version.