Remove the mime_size variable from imap_body_type_mpart_to_body, and pass 0 to mailmime_new instead.
The call to imap_body_fields_to_mime_fields passes NULL as the
body_fields, which meant that mime_size would not be set.
This meant in effect that it was 0 if no extended fields were set
(type_mpart->bd_ext_mpart == NULL) or uninitialized garbage otherwise.

This change sets it to 0 in all cases.