Don't free the struct mailimap_greeting after parsing the greeting. Instead, give it to the struct mailimap for safe keeping.
This means that client code can see the greeting (e.g. for debugging)
even after other commands have been issued.
This obviously consumes a bit of memory for the lifetime of the
session, but it's only a line of text and a couple of small structs
so I'm not worried about it.