Upgrade the embedded version of cyrus-sasl to 2.1.26.
This includes regeneration of the patch against lib/client.c, and
updates to prepare-cyrus-sasl.sh to specify the new version and
the new hosting website.

This doesn't include the upstream tar, even though 2.1.25 is committed
here, because I don't think that we should be doing that.  prepare-cyrus-sasl.sh
is capable of downloading it from upstream anyway.  This cset adds it
to .gitignore instead, and removes the 2.1.25 version.