Fix a memory leak when the response to an ID command returns NIL.
In that case, the code was allocating a params_list for the result, but
then not actually setting *result.  We got away with this (other than the
memory leak) only because mailimap_id_ext_data_free is tolerant of
the ext_data being NULL and mailimap_extension_data_new never failed.  If
either of those weren't true, we'd have crashed.